Move NullRemover into quasar.yggdrasil.util
