+++ yggdrasil/src/main/scala/quasar/yggdrasil/util/NullRemover.scala
@@ -1,0 +1,34 @@
+/*
+ * Copyright 2014–2018 SlamData Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package quasar.yggdrasil.util
+
+import quasar.precog.common.{ColumnRef, CNull}
+import quasar.yggdrasil.table.{Column, CScanner}
+
+object NullRemover extends CScanner {
+  type A = Unit
+  val init = ()
+
+  def scan(u: Unit, cols: Map[ColumnRef, Column], range: Range): (Unit, Map[ColumnRef, Column]) = {
+    val cols2 = cols flatMap {
+      case (ColumnRef(_, CNull), _) => Nil
+      case pair => List(pair)
+    }
+
+    ((), cols2)
+  }
+}
